refactor(AppInputForm): 禁用属性抽离为 Props 供外部传入, 进一步降低与对话场景的耦合
